[Geometry:Bugfix] fix more size_t->int narrowing (clang/MSVC hard errors)
Same class of clang -Wc++11-narrowing / MSVC C2398 errors as the previous
GeometryGather fix, in the remaining ops a full keep-going build surfaced:

- geometry/GeometryScatter.cpp: `auto N = indices->elementSize()` (size_t)
  feeds createDevice<int>({N}) / {N, D} and allocConst({N}); make N an int
  (buildScatterND already takes int N).
- backend/cpu/CPUROIAlign.cpp: createDevice<int32_t>({mROI.elementSize()})
  -- cast the size_t to int at the call site.

Verified a full libMNN.a build with -Werror=narrowing is now clean.

+++ source/backend/cpu/CPUROIAlign.cpp
@@ -40,7 +40,9 @@ ErrorCode CPUROIAlign::onResize(const std::vector<Tensor*>& inputs, const std::v
     TensorUtils::setLinearLayout(&mROI);
     auto core    = static_cast<CPUBackend*>(backend())->functions();
     if (core->bytes < 4) {
-        mROITemp.reset(MNN::Tensor::createDevice<int32_t>({mROI.elementSize()}));
+        // elementSize() is size_t; createDevice<int32_t>({...}) takes an
+        // initializer_list<int>, a narrowing conversion clang/MSVC reject.
+        mROITemp.reset(MNN::Tensor::createDevice<int32_t>({static_cast<int>(mROI.elementSize())}));
     }
     auto res = backend()->onAcquireBuffer(&mROI, Backend::DYNAMIC);
     if (!res) {
